refactor: rename events.importer.util to utils
Unify utility module naming

Refs: LINK-1675
